+++ packages/e2e/tests/ui-interactions.test.ts
@@ -1,7 +1,9 @@
+import { Element } from '@wdio/sync';
 import { makeLauncher } from '../../dashboard/rmf-launcher';
 import { login, overwriteClick } from './utils';
 
 describe('ui interactions', () => {
+  let doorAccordion: Element;
   const launcher = makeLauncher();
 
   before(async () => await launcher.launch());
@@ -14,6 +16,15 @@ describe('ui interactions', () => {
 
   function closeOmniPanel() {
     $(`#omnipanel [aria-label=Close]`).click();
+  }
+
+  function omniPanelMainMenu() {
+    $(`#omnipanel [aria-label=Back]`).click();
+  }
+
+  function openDoorAccordian() {
+    doorAccordion = $('.MuiAccordion-root*=main_door');
+    doorAccordion.click();
   }
 
   it('clicking a door on the map focuses it on the panel', () => {
@@ -42,5 +53,36 @@ describe('ui interactions', () => {
     robot.click();
 
     expect($(`.MuiAccordion-root*=${robotName}`).$('.MuiAccordionDetails-root')).toBeVisible();
+  });
+
+  it('Clicking on an equipment on the map resets the filter', () => {
+    omniPanelMainMenu();
+    $('[data-component=MainMenu] [data-item=Doors]').click();
+    // set value to filter input
+    $('[data-component=simple-filter]').$('input').setValue('value');
+
+    // get door marker and click
+    const door = $('[data-component=DoorMarker]');
+    const doorName = door.getAttribute('aria-label');
+
+    door.waitForClickable();
+    door.click();
+
+    // check that the filter is empty after clicking
+    expect($('[data-component=simple-filter]').$('input').getValue()).toEqual('');
+    // check that door panel is expanded
+    expect($(`.MuiAccordion-root*=${doorName}`).$('.MuiAccordionDetails-root')).toBeVisible();
+  });
+
+  it('clicking on open button opens the door', () => {
+    openDoorAccordian();
+    doorAccordion.$('button=Open').click();
+    expect(doorAccordion.$('[role=status]')).toHaveText('OPEN');
+  });
+
+  it('clicking on close button closes the door', () => {
+    openDoorAccordian();
+    doorAccordion.$('button=Close').click();
+    expect(doorAccordion.$('[role=status]')).toHaveText('CLOSED');
   });
 });
